store/tikv: exclude some tests for race detector.
